bump and refactor console_ui
