+++ src/rust/engine/ui/Cargo.toml
@@ -4,9 +4,6 @@ version = "0.0.1"
 edition = "2021"
 authors = ["Pants Build <pantsbuild@gmail.com>"]
 
-[lib]
-path = "src/console_ui.rs"
-
 [dependencies]
 console = { workspace = true }
 futures = { workspace = true }
